Moved the ParameterEncoding and Method logic into a separate Swift file.
